fix(tooltip): Attributes no longer evalued in isolate scope
The tooltip directive has been re-architected to use a new scope rather
than an isolate scope, but the templated tooltipPopup directive now has
an isolate scope to introduce a similar level of scope security. As the
isolate scope is gone, the transclusion logic has been removed. In
addition, tooltip attributes are $observed manually to assign the
necessary variables on the child scope needed by the tooltipPopup
directive.

Closes #78.